moving this functiond from scripts to cli
