+++ cg_lims/EPPs/files/base.py
@@ -6,6 +6,7 @@ from cg_lims.EPPs.files.file_to_udf import csv_well_to_udf
 from cg_lims.EPPs.files.hamilton.base import hamilton
 from cg_lims.EPPs.files.pooling_map.make_pooling_map import pool_map
 from cg_lims.EPPs.files.placement_map.make_96well_placement_map import placement_map
+from cg_lims.EPPs.files.csv_for_kapa_truble_shooting.csv_for_kapa_debug import trouble_shoot_kapa
 
 
 @click.group(invoke_without_command=True)
@@ -19,3 +20,4 @@ files.add_command(csv_well_to_udf)
 files.add_command(pool_map)
 files.add_command(placement_map)
 files.add_command(hamilton)
+files.add_command(trouble_shoot_kapa)
