+++ cg_lims/EPPs/files/csv_for_kapa_truble_shooting/csv_for_kapa_debug.py
@@ -1,0 +1,59 @@
+import csv
+import sys
+from typing import List
+
+import click
+import yaml
+from genologics.lims import Lims
+
+from cg_lims import options
+from cg_lims.EPPs.files.csv_for_kapa_truble_shooting.models import DebugKapaCSV, HEADERS
+from cg_lims.app.schemas.master_steps import (
+    HybridizeLibraryTWIST,
+    AliquotsamplesforenzymaticfragmentationTWIST,
+    KAPALibraryPreparation,
+    PoolsamplesforhybridizationTWIST,
+    CaptureandWashTWIST,
+    BeadPurificationTWIST,
+    BufferExchange,
+)
+from cg_lims.exceptions import LimsError
+
+
+def build_sample_row(lims: Lims, sample_id: str) -> list:
+    sample_row = DebugKapaCSV(SampleID=sample_id)
+    sample_row.set_hybridize(hybridize=HybridizeLibraryTWIST(lims=lims, sample_id=sample_id))
+    sample_row.set_aliquot(
+        aliquot=AliquotsamplesforenzymaticfragmentationTWIST(lims=lims, sample_id=sample_id)
+    )
+    sample_row.set_kapa(kapa=KAPALibraryPreparation(lims=lims, sample_id=sample_id))
+    sample_row.set_pool(pool=PoolsamplesforhybridizationTWIST(lims=lims, sample_id=sample_id))
+    sample_row.set_capture(capture=CaptureandWashTWIST(lims=lims, sample_id=sample_id))
+    sample_row.set_bead(bead=BeadPurificationTWIST(lims=lims, sample_id=sample_id))
+    sample_row.set_buffer(buffer=BufferExchange(lims=lims, sample_id=sample_id))
+
+    sample_row_dict = sample_row.dict(by_alias=True)
+    return [sample_row_dict.get(header) for header in HEADERS]
+
+
+@click.command()
+@options.samples_file(help="Txt file with sample ids")
+@options.file_placeholder(help="File placeholder name.")
+@click.pass_context
+def trouble_shoot_kapa(ctx, sample_file: str, file: str):
+    lims = ctx.obj["lims"]
+
+    try:
+        with open(sample_file, "r") as samples:
+            sample_list = [sample_id.strip("\n") for sample_id in samples.readlines()]
+
+        with open(f"{file}_kapa_debug.csv", "w", newline="\n") as new_csv:
+            wr = csv.writer(new_csv, delimiter=",")
+            wr.writerow(HEADERS)
+            sample_rows: List[List[str]] = [
+                build_sample_row(lims, sample_id) for sample_id in sample_list
+            ]
+            wr.writerows(sample_rows)
+        click.echo("Twist csv file has been generated.")
+    except LimsError as e:
+        sys.exit(e.message)
