+++ cg_lims/options.py
@@ -206,3 +206,11 @@ def concentration_udf_option(
     help: str = "String of UDF used to get concentration value",
 ) -> click.option:
     return click.option("--concentration-udf", required=True, help=help)
+
+
+def samples_file(help: str = "Txt file with sample ids") -> click.option:
+    return click.option(
+        "--samples-file",
+        required=True,
+        help=help,
+    )
